Crate elastic_hyper

Elasticsearch Hyper Client

A lightweight implementation of the Elasticsearch API based on Hyper.

Each API endpoint is represented as its own function, so each possible http route gets its own function. The functions are also designed to work well with the elastic_types and json_str crates, but deserialisation is the responsibility of the caller.

Usage

This crate is on crates.io. To get started, add elastic_hyper and hyper to your Cargo.toml:

[dependencies]
elastic_hyper = "*"
hyper = "*"

For Windows, you may need to exclude openssl or the build can fail:

[dependencies]
elastic_hyper = { version = "*", default-features = false }
hyper = { version = "*", default-features = false }

And reference them in your crate root:

extern crate elastic_hyper as elastic;
extern crate hyper;

Ping the availability of your cluster:

let (mut client, params) = elastic::default();

//HTTP HEAD /
elastic::ping::head(&mut client, &params).unwrap();

Ping the availability of your cluster:

let mut client = hyper::Client::new();
let mut params = elastic::RequestParams::default()
    .url_params(vec![
        ("q", "'my string'".to_owned())
    ]);

//HTTP GET /myindex/mytype/_search?q='my string'
elastic::search::get_index_type(&mut client, &params, "myindex", "mytype").unwrap();

Links

elastic_types

json_str

Github

Modules

bulk
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-bulk.html

cat

clear_scroll
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-request-scroll.html

cluster

count
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-count.html

count_percolate
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-percolate.html

delete
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-delete.html

delete_script
http://www.elastic.co/guide/en/elasticsearch/reference/master/modules-scripting.html

delete_template
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-template.html

exists
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-get.html

explain
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-explain.html

field_stats
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-field-stats.html

get
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-get.html

get_script
http://www.elastic.co/guide/en/elasticsearch/reference/master/modules-scripting.html

get_source
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-get.html

get_template
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-template.html

index
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-index_.html

indices

info
http://www.elastic.co/guide/

ingest

mget
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-multi-get.html

mpercolate
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-percolate.html

msearch
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-multi-search.html

mtermvectors
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-multi-termvectors.html

nodes

percolate
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-percolate.html

ping
http://www.elastic.co/guide/

put_script
http://www.elastic.co/guide/en/elasticsearch/reference/master/modules-scripting.html

put_template
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-template.html

reindex
https://www.elastic.co/guide/en/elasticsearch/plugins/master/plugins-reindex.html

render_search_template
http://www.elasticsearch.org/guide/en/elasticsearch/reference/master/search-template.html

scroll
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-request-scroll.html

search
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-search.html

search_shards
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-shards.html

search_template
http://www.elastic.co/guide/en/elasticsearch/reference/current/search-template.html

snapshot

suggest
http://www.elastic.co/guide/en/elasticsearch/reference/master/search-suggesters.html

tasks

termvectors
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-termvectors.html

update
http://www.elastic.co/guide/en/elasticsearch/reference/master/docs-update.html

update_by_query
https://www.elastic.co/guide/en/elasticsearch/plugins/master/plugins-reindex.html

Structs

RequestParams
Misc parameters for any request.

Functions

default
Get a default Client and RequestParams.