Struct elastic_hyper::RequestParams

pub struct RequestParams {
    pub base_url: String,
    pub url_params: BTreeMap<&'static str, String>,
    pub headers: Headers,
}

Misc parameters for any request.

The RequestParams struct allows you to set headers and url parameters for your requests. By default, the ContentType::json header will always be added. Url parameters are added as simple key-value pairs, and serialised by rust-url.

Examples

With default query parameters:

extern crate hyper;
extern crate elastic_hyper as elastic;

let params = elastic::RequestParams::default();

With custom headers:

extern crate hyper;
extern crate elastic_hyper as elastic;

let mut params = elastic::RequestParams::default();

//Add your own headers
params.headers.set(hyper::header::Authorization("let me in".to_owned()));

Add url query parameters to the request:

extern crate hyper;
extern crate elastic_hyper as elastic;

let params = elastic::RequestParams::default()
        .url_params(vec![
            ("pretty", "true".to_owned()),
            ("q", "*".to_owned())
        ]);

With a custom base url:

extern crate hyper;
extern crate elastic_hyper as elastic;

let params = elastic::RequestParams::new("http://mybaseurl:9200", hyper::header::Headers::new());

Fields
base_url: String

Base url for Elasticsearch

url_params: BTreeMap<&'static str, String>

Simple key-value store for url query params.

headers: Headers

The complete set of headers that will be sent with the request.

Methods

impl RequestParams

fn new<T: Into<String>>(base: T, headers: Headers) -> Self

Create a new container for request parameters.

Attempts to add ContentType::json to the passed in headers param.

fn url_params<I>(self, url_params: I) -> Self where I: IntoIterator<Item=(&'static str, String)>

Add a collection of url params.

fn get_url_qry(&self) -> String

Get the url params as a formatted string.

Follows the application/x-www-form-urlencoded format.